requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.2
pandas_ta==0.3.14b0
numpy==2.1.1
groq==0.11.0
